+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexValueMixin.kt
@@ -19,17 +19,26 @@
 package com.intellij.idea.plugin.hybris.impex.psi.impl
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexFullHeaderParameter
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexValue
+import com.intellij.idea.plugin.hybris.impex.psi.references.ImpExTSComposedTypeValueReference
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpExTSDynamicEnumValueReference
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpExTSStaticEnumValueReference
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexTSAttributeReference
 import com.intellij.idea.plugin.hybris.system.type.meta.TSMetaModelAccess
+import com.intellij.idea.plugin.hybris.system.type.meta.TSModificationTracker
+import com.intellij.idea.plugin.hybris.system.type.meta.model.TSGlobalMetaEnum
+import com.intellij.idea.plugin.hybris.system.type.meta.model.TSGlobalMetaItem
 import com.intellij.idea.plugin.hybris.system.type.psi.reference.result.AttributeResolveResult
 import com.intellij.lang.ASTNode
+import com.intellij.openapi.components.service
 import com.intellij.psi.*
+import com.intellij.psi.util.CachedValueProvider
+import com.intellij.psi.util.CachedValuesManager
+import com.intellij.psi.util.PsiModificationTracker
 import com.intellij.sql.indexOf
 import com.intellij.util.asSafely
 import java.io.Serial
@@ -45,7 +54,14 @@ abstract class ImpexValueMixin(node: ASTNode) : ASTWrapperPsiElement(node), PsiL
 
     override fun getReference() = references.firstOrNull()
 
-    override fun getReferences(): Array<out PsiReference?> {
+    override fun getReferences(): Array<PsiReference> = CachedValuesManager.getManager(project).getCachedValue(this) {
+        CachedValueProvider.Result.create(
+            calculateReferences(),
+            project.service<TSModificationTracker>(), PsiModificationTracker.MODIFICATION_COUNT
+        )
+    }
+
+    private fun calculateReferences(): Array<PsiReference> {
         val hasMacroUsage = children
             .filterIsInstance<ImpexValue>()
             .any { it.children.any { subChild -> subChild is ImpexMacroUsageDec } }
@@ -63,20 +79,33 @@ abstract class ImpexValueMixin(node: ASTNode) : ASTWrapperPsiElement(node), PsiL
             ?.asSafely<AttributeResolveResult>()
             ?.meta
             ?.type
-            ?.let { TSMetaModelAccess.getInstance(project).findMetaEnumByName(it) }
+            ?.let { TSMetaModelAccess.getInstance(project).findMetaClassifierByName(it) }
             ?.let { meta ->
-                val index = getEnumCodeParameterIndex(fullHeaderParameter)
-                    ?.takeIf { getFieldValues().size > it }
-                    ?: return emptyArray()
+                val name = meta.name ?: return emptyArray()
 
-                if (meta.isDynamic) ImpExTSDynamicEnumValueReference(this, index, meta.name!!)
-                else ImpExTSStaticEnumValueReference(this, index, meta.name!!)
+                when {
+                    meta is TSGlobalMetaEnum -> {
+                        val index = getParameterIndex(fullHeaderParameter) ?: return@let null
+
+                        if (meta.isDynamic) ImpExTSDynamicEnumValueReference(this, index, name)
+                        else ImpExTSStaticEnumValueReference(this, index, name)
+                    }
+
+                    meta is TSGlobalMetaItem && HybrisConstants.TS_COMPOSED_TYPE == name -> {
+                        val index = getParameterIndex(fullHeaderParameter) ?: return@let null
+
+                        ImpExTSComposedTypeValueReference(this, index, name)
+                    }
+
+                    else -> null
+                }
+
             }
-            ?.let { arrayOf<PsiReference>(it) }
+            ?.let { arrayOf(it) }
             ?: emptyArray()
     }
 
-    private fun getEnumCodeParameterIndex(fullHeaderParameter: ImpexFullHeaderParameter) = fullHeaderParameter.parametersList
+    private fun getParameterIndex(fullHeaderParameter: ImpexFullHeaderParameter, attributeName: String = "code") = fullHeaderParameter.parametersList
         .firstOrNull()
         ?.parameterList
         ?.indexOf { parameter ->
@@ -87,9 +116,10 @@ abstract class ImpexValueMixin(node: ASTNode) : ASTWrapperPsiElement(node), PsiL
                 ?.asSafely<AttributeResolveResult>()
                 ?.meta
                 ?.name
-                ?.let { it == "code" }
+                ?.let { it == attributeName }
                 ?: false
         }
+        ?.takeIf { it >= 0 && getFieldValues().size > it }
 
     private fun getFieldValues(): Array<PsiElement> = findChildrenByType(ImpexTypes.FIELD_VALUE, PsiElement::class.java)