+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 ### `ImpEx` enhancements
 - Enhanced Enum values reference resolution within Enum column [#1360](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1360)
+- Added Item values reference resolution within ComposedType attribute columns [#1361](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1361)
 
 ### `Terminal` integration
 - Added support of the `Terminal` plugin and set default directory to the project root for new Terminal Windows [#1342](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1342)